Created Class page, Edit User page and changed visuals of User Details
Added Class.vue, which is meant to display the currently selected class and the enrolled students as well as who is attended. Added Edit User Details page for teachers and admins to be able to change individuals details on the system. Added routes for new pages. Changed View User to be more aesthetically pleasing